% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleancall.R
\name{call_with_cleanup}
\alias{call_with_cleanup}
\title{Call a native routine within an exit context}
\usage{
call_with_cleanup(ptr, ...)
}
\arguments{
\item{ptr}{A native pointer object.}

\item{...}{Arguments for the native routine.}
}
\description{
Call a native routine within an exit context
}
