% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\docType{package}
\name{cleancall-package}
\alias{cleancall}
\alias{cleancall-package}
\title{cleancall: C Resource Cleanup via Exit Handlers}
\description{
Wrapper of .Call() that runs exit handlers to clean up C resources. Helps managing C (non-R) resources while using the R API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/cleancall#readme}
  \item \url{https://r-lib.github.io/cleancall/}
  \item Report bugs at \url{https://github.com/r-lib/cleancall/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com} (\href{https://orcid.org/0000-0001-7098-9676}{ORCID})

Authors:
\itemize{
  \item Lionel Henry \email{lionel@posit.co}
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
