% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pushexit.R
\name{ptr_invoke}
\alias{ptr_invoke}
\title{Call a native routine within an exit context}
\usage{
ptr_invoke(ptr, ...)
}
\arguments{
\item{ptr}{A native pointer object.}

\item{...}{Arguments for the native routine.}
}
\description{
Call a native routine within an exit context
}
