% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleancall.R
\docType{package}
\name{cleancall-package}
\alias{cleancall}
\alias{cleancall-package}
\title{cleancall: Push exit handlers for C resources}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@rstudio.com}

Authors:
\itemize{
  \item Gábor Csárdi \email{csardi.gabor@gmail.com} (0000-0001-7098-9676)
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
