% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleancall.R
\docType{package}
\name{cleancall-package}
\alias{cleancall}
\alias{cleancall-package}
\title{cleancall: C Resource Cleanup via Exit Handlers}
\description{
Wrapper of '.Call()' that runs exit handlers to clean up
    C resources. Helps when managing C (non-R) resources while using the
    R API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/cleancall#readme}
  \item Report bugs at \url{https://github.com/r-lib/cleancall/issues}
}

}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@rstudio.com}

Authors:
\itemize{
  \item Gábor Csárdi \email{csardi.gabor@gmail.com} (0000-0001-7098-9676)
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
